ES/mapping: add task parameters to progress mapping.
To avoid extraction of task IDs for every campaign step and then
querying progress statistics by these IDs, the parameters that allows to
detect task's affilation with a campaign and its step are added to the
progress info documents.

In theory, a campaign can be said by a hashtag and a step -- by a
combination of AMI tags chain and output data format.

For now, however, we use MC steps -- or the last (current) AMI tag with
format. So all three kinds of "steps" are added to the mapping.