DF/progress4es/020: allow skipping "wrong" messages.
If an input message is somehow "incorrect" (e.g. has no "taskid" field)
or no information was found in the DKB ES -- the input message will be
passed through as-is -- for it is not up to this stage to filter
something out, right? However it can mark a message as "incomplete",
just in case.