ES/mapping: rename 'hashtag' to 'hashtage_list' (progress data).
In the mapping for tasks metadata we already have "hashtag_list", and
supposedly these fields are to be used in similar queries -- so naming
them differently would only complicate request logic.

Althought "hashtag_list" does not look very good, but -- let's keep
things consistent. Should we change the naming -- we'll change it in
both indices simultaneously.