DF/009: change query configuration structure.
This change makes the config file less dependent on the stage source
directory location.

Queries themselves can be quite complicated, so usually they are
considered to be more a part of the stage itself, rather than its
configuration. So whenever the stage directory moved (e.g. the
repository is checked out to different location), the configuration
becomes invalid and it is not good.

Now instead of the full (or relative to the configuration file) path to
the queries sources, one is supposed to configure query names and type
-- and the stage will look for them in the stage source directory:

  $STAGE_DIR/query[/$query_type]/$query_name

The `$query_type` parameter is optional and is introduced for cases when
the single stage is used in different scenarios (here: `data4es`
integration process and process for consostency check of the integrated
by this process data).

NOTE: currently the stage implementation hardly relays on the query
names ('tasks' and 'datasets'), so changes in query names should be done
with caution.