DF/data4es: add (possible) step names to task metadata.
New fields are added on stage 017 (adjustMetadata):
* ctag_format_step;
* ami_tags_format_step.

There are different ways to say one step from another:
* MC production step name (already exists as 'step_name' field);
* current AMI tag + output data format;
* chain of AMI tags + output data format.

The latter is supposed to be the most universal, but initially only the
first one was used, then for some cases the second was invented, and...
...and there's no good way to make things as they are supposed to be all
at once. So we need all the possible namings.

---

NOTE.
Step names are based on the output data formats. They can be taken from
task's metadata `output_formats` field, but it wasn't used before (when
we needed tasks classification by format-based-steps). However,
according to M.Borodin, field 'output_formats' and list of data formats,
derived from the list of dataset names, must be exactly the same (if
format is specified in 'output_formats', the task _will_ produce a
dataset in the given format).