DF/data4es: add (possible) step names to task metadata.
New fields are added on stage 017 (adjustMetadata):
* ctag_format_step;
* ami_tags_format_step.

There are different ways to say one step from another:
* MC production step name (already exists as 'step_name' field);
* current AMI tag + output data format;
* chain of AMI tags + output data format.

The latter is supposed to be the most universal, but initially was used
only the first one, then for some cases the second was invened, and...
...and there's no good way to make things as they are supposed to be all
at once. So we need all the possible namings.

---

NOTE 1.
Step names are based on the output data formats. They can be taken from
task's metadata `output_formats` field, but it wasn't used before (when
we needed tasks classification by format-based-steps), so I am not sure
can or cannot it be used here.

NOTE 2.
Since data format is taken not from `output_formats`, but from output
datasets' names, the functionality from Stage 093 (datasetFormats) is
reused by moving it into `pyDKB` library (module `atlas.misc` is
created for this kind of more-than-once used functions).