DF/data4es/017: add calculation of {input|processed}_events_v2.
New algorithm of calculation for these values was suggested by
M.Borodin <Mikhail.Borodin@cern.ch> during a discussion on "how to get
current progress info for a task from DEFT" as the "most general" one.

There can possibly be some cases not addressed in the current version,
but they are to be detected when we start collecting and using the data.