DF/data4es: add separate field for AMI tags chain.
We have `ctag` field with the current (last) AMI tag; but fully
understand the process, to which the given task belongs, we can only by
the whole chain of AMI tags (as they basically define the steps of data
processing).

This field is not required now for any use-case, but will allow more
accurate classification of tasks by steps and, consequently, let us
provide more adequate results in some use-cases.