DF/data4es/093: refactoring.
1. Extracting a specific field from a dataset name is a single logical
operation -- so it is much more readable to have a function call "give
me data format" then the whole algorithm's implementation within a more
high-level operation (in this case, getting a list of formats that
should be associated to the dataset).

2. Dataset name parsing doesn't look like a stage-specific functionality,
more like a whole use-case specific one. So it is moved to the (new)
case-specific module of the common library.