DF/data4es: adapt process for nested ES scheme (pt.1: Stage 091).
The new ES indexing scheme (mapping) will expect all metadata (of a task
and its output datasets) to be passed in a single document, where what
previously was a number of individual documents of 'output_dataset' type
becomes a list of nested documents:
```
{
  _id: <taskid>,
  ...<task_metadata>...,
  output_dataset: [
    {name: ..., bytes: ..., ...},
    ...
  ]
}
```

Stage 091 was used to divide the flow of messages in two: with task
metadata (which also contained the primary input dataset metadata) and
output dataset metadata (linked to the task via '_parent: <taskid>'
field).

New version keeps the flow of messages undivided, passing to the
downstream stages messages corresponding the new format (with nested
'output_dataset' documents).