ES/mapping: clone tasks mapping to turn into nested scheme.
Switching to a new storage scheme involves a lot of changes, including:
* mapping update;
* DF modification;
* reindexing of ES data;
* running the new DF;
* API modification.

It would be hard to do all the things at once, so the nested version
will first be added as an independent instance -- with its own scheme,
integration process and API methods. It will be running in parallel with
the already operating services, and when all the necessary preparations
are done and we have the fully functional DKB with the new scheme --
current version will be disabled.

That's the plan, at least ;)