Set EOB to be non-empty by default.
Default arguments are supposed to make things work as expected - while
EOB's value does not really matter now, without a supervisor, it would
matter with it.

Existing stages are not affected because all markers (which currently consist
only of EOB) are ignored by InputStream when batch mode is not used.