DF/data4es: fix Stage 19,69's input/output file names in debug mode.
There are two "sink chains" in the process with different data streams;
so in debug mode we need to create different files for stages in
different chains (like '69-b40.out' and '69-b91.out' instead of simple
'69.out').

For `data4es-nested` it is not necessary, but since the "sink chain" is
defined as a function -- it _can_ be called more than once, so suffixes
to debug files should also be added.