API: add '/nested/' endpoint.
For each method called via this endpoint, e.g. like this:

  http://api.atlas-dkb.cern.ch:5080/nested/task/chan?tid=16655409

...is to be used its alternative 'nested' implementation. If not
defined, or not implemented yet -- default implementation will be used,
and appropriate warning message will be added to the response metadata.

---

Alternative implementation is defined as a new instance of
`DKBStorageMethod` and linked to the default implementation via
the default one's `alts` property.

To make use of the alt method, the default one is called with `_alt`
parameter (its value is the name of the alternative, in this case:
'nested').

It might be better to have alts defined at the handlers' level as well
(or instead?) -- to avoid icounter-intuitive "special"
parameters-related stuff) -- but it is a bit too much for now...