API: fix issue with "any category" regex path standardizing.
Method handler added to "any category" ended up linked to a regex
`/.*/metod_name`, to which simple `/metod_name` wouldn't match; so in
such cases leading `/` is not to be added artificially. But in cases
where it _is_ required before "any symbol", it still can (and should) be
specified explicitly.