DF/data4es(-nested)/095: improve try/except logic.
The operation `change_key_names()` makes no sence in case when there was
any exception; so placing it after `try/except` clause just "as is"
looks a bit wierd.

Placing it into `else` section of `try/except` says: "and if
everything's fine we need to do this thing". This looks better.