ES/reindex: fix NOOP mode (failed to save last TID).
When we have a pipeline like this:

  cat file | tee some_file | echo 1

the `tee` operation has it's STDOUT connected to `echo`'s STDIN, which
closes too soon -- and `tee` fails to write to STDOUT (exit code: 141).
And as a result, it doesn't write to `some_file` as well -- so it's
empty. And when `reindex.sh` tries to save last TID, it fails to find
it in the empty file.